Move savedQuery UI to separate file
